Move i18n init
